uop: move BufferizeOpts to ops, breaking codegen-schedule cycle
